cleaned up tests
