+++ tests/unit/test_synapse_fs_unit.py
@@ -93,3 +93,23 @@ class TestSynapseErrors:
         with pytest.raises(SynapseHTTPError):
             with synapse_errors("foo"):
                 raise SynapseHTTPError("something else")
+
+
+class TestPathToParentId:
+    """Tests for SynapseFS._path_to_parent_id edge cases."""
+
+    def test_rootless_bare_name_raises_value_error(self) -> None:
+        """Verify that a bare name in rootless mode raises ValueError."""
+        fs = SynapseFS()
+        with pytest.raises(ValueError, match="must start with a Synapse ID"):
+            fs._path_to_parent_id("not_a_synapse_id")
+
+
+class TestPathToSynapseId:
+    """Tests for SynapseFS._path_to_synapse_id edge cases."""
+
+    def test_rootless_non_id_path_raises_value_error(self) -> None:
+        """Verify that a path not starting with a Synapse ID raises ValueError."""
+        fs = SynapseFS()
+        with pytest.raises(ValueError):
+            fs._path_to_synapse_id("SynapseFS Test Project/syn50555279")
